Update float16 type-casting.

+++ src/plugins/intel_gpu/tests/test_utils/test_utils.h
@@ -616,7 +616,7 @@ inline cldnn::memory::ptr get_generated_random_1d_mem(cldnn::engine& engine, cld
         VF<uint8_t> rnd_vec = generate_random_1d<uint8_t>(s.count(), -200, 200);
         set_values(prim, rnd_vec);
     } else if (l.data_type == cldnn::data_types::f16) {
-        VF<uint16_t> rnd_vec = generate_random_1d<uint16_t>(s.count(), -1, 1);
+        VF<FLOAT16> rnd_vec = generate_random_1d<FLOAT16>(s.count(), -1, 1);
         set_values(prim, rnd_vec);
     } else {
         VF<float> rnd_vec = generate_random_1d<float>(s.count(), -1, 1);
